transcription_service/pyproject.toml
[tool.poetry]
name = "transcription-service"
version = "0.1.0"
description = "Transcription Service with Whisper"
authors = ["usnar"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "*"
requests = "*"
uvicorn = "*"
redis = "*"
celery = "*"
python-dotenv = "*"
aiohttp = "3.12.13"
openai-whisper = {git = "https://github.com/openai/whisper.git"}

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-asyncio = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"